chore: Move models to models instead of repos
