+++ src/domains/collection/models.py
@@ -1,6 +1,6 @@
 from django.db import models
 
 # Create your models here.
-from domains.collection.infra.repos.user import User
-from domains.collection.infra.repos.collector import Collector
-from domains.collection.infra.repos.task import Task
+from domains.collection.infra.models.user import User
+from domains.collection.infra.models.collector import Collector
+from domains.collection.infra.models.task import Task
